fix: circ deps
