+++ src/domain/shared/error-parsers.ts
@@ -1,5 +1,3 @@
-import { UnknownDomainError } from "./errors"
-
 export const parseErrorMessageFromUnknown = (error: unknown): string => {
   const errMsg =
     error instanceof Error
@@ -21,17 +19,5 @@ export const parseErrorFromUnknown = (error: unknown): Error => {
       : error instanceof Object
       ? new Error(JSON.stringify(error))
       : new Error("Unknown error")
-  return err
-}
-
-export const parseUnknownDomainErrorFromUnknown = (error: unknown): DomainError => {
-  const err =
-    error instanceof Error
-      ? error
-      : typeof error === "string"
-      ? new UnknownDomainError(error)
-      : error instanceof Object
-      ? new UnknownDomainError(JSON.stringify(error))
-      : new UnknownDomainError("Unknown error")
   return err
 }
