+++ src/domain/shared/index.ts
@@ -6,6 +6,7 @@ export * from "./amount"
 export * from "./safe"
 export * from "./errors"
 export * from "./error-parsers"
+export * from "./error-parsers-unknown"
 
 export const setErrorWarn = (error: DomainError): DomainError => {
   error.level = ErrorLevel.Warn
